extract conflicted file component
